Deployment/requirements.txt
waitress
numpy
Pillow
streamlit
tensorflow-cpu
